+++ crates/jsonschema/src/error.rs
@@ -36,11 +36,9 @@ pub struct ValidationError<'a> {
 /// let schema = json!({"maxLength": 5});
 /// let instance = json!("foo");
 /// if let Ok(validator) = jsonschema::validator_for(&schema) {
-///     let result = validator.validate(&instance);
-///     if let Err(errors) = result {
-///         for error in errors {
-///             println!("Validation error: {}", error)
-///         }
+///     let errors = validator.iter_errors(&instance);
+///     for error in errors {
+///         println!("Validation error: {}", error)
 ///     }
 /// }
 /// ```
@@ -998,7 +996,7 @@ mod tests {
             }
         );
         let validator = crate::validator_for(&schema).unwrap();
-        let mut result = validator.validate(instance).expect_err("error iterator");
+        let mut result = validator.iter_errors(instance);
         let error = result.next().expect("validation error");
 
         assert!(result.next().is_none());
@@ -1039,7 +1037,7 @@ mod tests {
             }
         );
         let validator = crate::validator_for(&schema).unwrap();
-        let mut result = validator.validate(instance).expect_err("error iterator");
+        let mut result = validator.iter_errors(instance);
         let error = result.next().expect("validation error");
 
         assert!(result.next().is_none());
@@ -1064,7 +1062,7 @@ mod tests {
             }
         );
         let validator = crate::validator_for(&schema).unwrap();
-        let mut result = validator.validate(instance).expect_err("error iterator");
+        let mut result = validator.iter_errors(instance);
         let error = result.next().expect("validation error");
 
         assert!(result.next().is_none());
@@ -1086,7 +1084,7 @@ mod tests {
             }
         );
         let validator = crate::validator_for(&schema).unwrap();
-        let mut result = validator.validate(instance).expect_err("error iterator");
+        let mut result = validator.iter_errors(instance);
         let error = result.next().expect("validation error");
 
         assert!(result.next().is_none());