+++ crates/jsonschema/src/keywords/custom.rs
@@ -1,7 +1,7 @@
 use crate::{
     paths::{LazyLocation, Location},
     validator::Validate,
-    ErrorIterator, ValidationError,
+    ValidationError,
 };
 use serde_json::{Map, Value};
 
@@ -16,7 +16,11 @@ impl CustomKeyword {
 }
 
 impl Validate for CustomKeyword {
-    fn validate<'i>(&self, instance: &'i Value, location: &LazyLocation) -> ErrorIterator<'i> {
+    fn validate<'i>(
+        &self,
+        instance: &'i Value,
+        location: &LazyLocation,
+    ) -> Result<(), ValidationError<'i>> {
         self.inner.validate(instance, location)
     }
 
@@ -33,7 +37,11 @@ pub trait Keyword: Send + Sync {
     /// easily or efficiently expressed in JSON schema.
     ///
     /// The custom validation is applied in addition to the JSON schema validation.
-    fn validate<'i>(&self, instance: &'i Value, location: &LazyLocation) -> ErrorIterator<'i>;
+    fn validate<'i>(
+        &self,
+        instance: &'i Value,
+        location: &LazyLocation,
+    ) -> Result<(), ValidationError<'i>>;
     /// Validate instance and return a boolean result.
     ///
     /// Could be potentilly faster than [`Keyword::validate`] method.